APP-33: Extract formatRelativeTime to app/lib/relative-time for shared use.
